PEP8: Spacing around comments
Comments should start with `# ` and should have at least 2 spaces if they are an
inline comment.

I have chosen not to change the spacing before commented out code as that makes
it less clear what is a normal comment vs code when skimming through files. I
also plan on removing any commented out code in a separate patch anyway so the
issue will be moot.