Don't use format() in log statements.
The logging convenience functions logging.debug(), logging.info() etc. accept a
variable number of arguments with formatting values for the format string
passed as first argument. Formatting the string is deferred until it is
verified that it is really needed, i.e. logging for level DEBUG is enabled when
calling logging.debug().

socketcan_ctypes.py contained a lot of statements on the form:

    log.debug("Format string with argument: {}".format(value))

The format() function was called each time the statement was run, _even though
logging for level DEBUG was enabled_. Since this happened in the code that is
run for each CAN frame, it cost a lot of processing time.

I changed each log point to %-style formatting instead of {}-style. This is
simply because Python 2 does not support any other type of formatting in the
logging convenience functions.