Unified timeout behavior between socketcan_native and socketcan_ctypes.
I've changed the socketcan_native timeout behavior to act in the same way as
socketcan_ctypes. So, it now makes use of the select.select() function instead
of socket.settimeout().

I've also update the documentation to reflect this behavior. Now, the
documentation of socketcan reading/timeout behavior is common between both
socketcan_native and socketcan_ctypes.